3dobject: fix warnings.

+++ game/3dobject.cc
@@ -89,9 +89,9 @@ void Object3d::loadWavefrontObj(fs::path const& filepath, float scale) {
 		bool hasNormals = !i->normals.empty();
 		bool hasTexCoords = !i->texcoords.empty();
 		for (size_t j = 0; j < i->vertices.size(); ++j) {
-			if (hasNormals) m_va.normal(normals[i->normals[j]]);
-			if (hasTexCoords) m_va.texCoord(texcoords[i->texcoords[j]]);
-			m_va.vertex(vertices[i->vertices[j]]);
+			if (hasNormals) m_va.normal(normals[static_cast<size_t>(i->normals[j])]);
+			if (hasTexCoords) m_va.texCoord(texcoords[static_cast<size_t>(i->texcoords[j])]);
+			m_va.vertex(vertices[static_cast<size_t>(i->vertices[j])]);
 		}
 	}
 }
